amidstexporter: Copy from amidst PKGBUILD.
AMIDST is currently not properly maintained, due to @skiphs being RL busy. A
couple of notable forks have sprung up, this one by @Treer and one by @dtfinch.
@Treer's branch is the one more often referenced in the AMIDST thread on
the MC forums (and the only one of the two to have proper releases on
GitHub), so that is the one I decided to package.

@Treer's motivation for his fork is given in this post:
http://www.minecraftforum.net/forums/mapping-and-modding/minecraft-tools/1262200-v3-7-amidst-strongholds-village-biome-etc-finder?comment=4719